Merge config object and use variable curve string
This commit resolves an issue where a chain of of properties throws:

    TypeError: Cannot read property 'incoming' of undefined

Using `_.merge()` instead merges the two objects, which makes changes
to nested objects without any errors. This commit also replaces the
hardcoded "ed25519" curve with a variable, which should be tolerant of
new curves that may be used in the future.